Replace allow_dependent_rows and allow_dependent_columns with assume_full_rank (#158)
The two functions allow_dependent_{rows,columns} together did the job of
answering if the solver accepts full rank matrices for the purposes of
the jvp. Allowing them to be implemented separately created some issues:
1) Invalid states were representable. Eg. What does it mean that
   dependent columns are allowed for square matrices if dependent rows
   are not? What does it mean that dependent rows are not allowed for
   matrices with more rows than columns?
2) As the functions accept operator as input, a custom solver could in
   principle decide its answer based on operator's dynamic value rather
   than only jax compilation static information regarding it, as in all
   the lineax defined solvers. This would prevent jax compilation and
   jit.

Both issues are addressed by asking the solver to report only if it
assumes the input is numerically full rank. If this assumption is
exactly violated, its behavior is allowed to be undefined, and is
allowed to error, produce NaN values, and produce invalid values.